requirements.txt
google-cloud-pubsub==2.11.0
elasticsearch==7.10.1
